# backend/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Boleteria Online",
    "description": "Colleción de API's correspondientes al sistema de Pasajes Online.",
    "contact": {
      "name": "Agustin Ibañez",
      "email": "admin@boleteriasystem.com"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/nodos": {
      "get": {
        "tags": [
          "Nodos"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/NodoResponseICollectionWebResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodoResponseICollectionWebResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodoResponseICollectionWebResult"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Nodos"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NodoRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/NodoRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/NodoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/NodoResponseWebResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodoResponseWebResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodoResponseWebResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "201": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/NodoResponseWebResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodoResponseWebResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodoResponseWebResult"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Nodos"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NodoRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/NodoRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/NodoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/NodoResponseWebResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodoResponseWebResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodoResponseWebResult"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Nodos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/NodoResponseWebResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodoResponseWebResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodoResponseWebResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/nodos/{id}": {
      "get": {
        "tags": [
          "Nodos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/NodoResponseWebResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodoResponseWebResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodoResponseWebResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DestinoResponse": {
        "required": [
          "nombre"
        ],
        "type": "object",
        "properties": {
          "nombre": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "ErrorMessage": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6,
          7,
          8,
          9,
          10,
          11,
          12
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "NodoRequest": {
        "required": [
          "demora",
          "destino_id",
          "origen_id",
          "precio"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "origen_id": {
            "type": "integer",
            "format": "int32"
          },
          "destino_id": {
            "type": "integer",
            "format": "int32"
          },
          "demora": {
            "type": "string",
            "format": "time"
          },
          "precio": {
            "type": "number",
            "format": "float"
          }
        },
        "additionalProperties": false
      },
      "NodoResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "origen": {
            "$ref": "#/components/schemas/DestinoResponse"
          },
          "destino": {
            "$ref": "#/components/schemas/DestinoResponse"
          },
          "demora": {
            "type": "string",
            "nullable": true
          },
          "precio": {
            "type": "number",
            "format": "float"
          }
        },
        "additionalProperties": false
      },
      "NodoResponseICollectionWebResult": {
        "type": "object",
        "properties": {
          "result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NodoResponse"
            },
            "nullable": true
          },
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "error_messages": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          },
          "error_code": {
            "$ref": "#/components/schemas/ErrorMessage"
          },
          "error": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "NodoResponseWebResult": {
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/NodoResponse"
          },
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "error_messages": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          },
          "error_code": {
            "$ref": "#/components/schemas/ErrorMessage"
          },
          "error": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  }
}